perf(brep): instrument the decoder CartesianPoint cache + add hoist primitives
Add point_cache_hits/misses counters to EntityDecoder (bumped in
get_polyloop_coords_cached) plus take_point_cache/set_point_cache/
point_cache_stats so the cache can be moved across per-element decoders.
Extract the decoder unit tests into a ratchet-exempt decoder/decoder_tests.rs
sibling and add two guards: cross-loop memoization fires (hits>0) and the
cache survives take->set across decoder instances.
